Fix directory renames
